Added common base class for Workers and "isWalkable(WalkPosition)"
